requirements.txt
django
Pillow
flake8
django-bootstrap5
django-filter